Add logic to prevent duplicating songs in queue
