+++ src/pages/room-page/components/playlists/playlists.tsx
@@ -1,19 +1,14 @@
 import React, { useCallback, useEffect, useState } from "react";
-import { useSelector } from "react-redux";
-import { useParams } from "react-router-dom";
 import { useMusic } from "lib/music/hook";
 import { Playlist } from "lib/music/types";
 import { MediaCard } from "components/media-card/media-card.component";
 import { MediaQueue, MediaQueueItem } from "components/media-queue";
-import { useFirebase } from "lib/firebase/hook";
-import { RootState } from "store/reducers";
+import { useEnqueueSongFirebase } from "hooks/use-enqueue-song-firebase";
 
 export const Playlists = () => {
   const music = useMusic();
-  const firebase = useFirebase();
-  const { roomKey } = useParams<{ roomKey: string }>();
-  const userId = useSelector((state: RootState) => state.userId);
   const [playlists, setPlaylists] = useState<Playlist[]>([]);
+  const enqueueSongFirebase = useEnqueueSongFirebase();
 
   useEffect(() => {
     music.getPlaylists().then((p) => setPlaylists(p));
@@ -24,10 +19,10 @@ export const Playlists = () => {
       const songs = await music.getSongsForPlaylist(playlist);
 
       songs.forEach((song) => {
-        firebase.database().ref(`rooms/${roomKey}/queues/${userId}`).push(song);
+        enqueueSongFirebase(song);
       });
     },
-    [firebase, roomKey, userId]
+    [music, enqueueSongFirebase]
   );
 
   return (
