+++ src/pages/room-page/components/search/search.tsx
@@ -1,15 +1,12 @@
 import React, { useState, useEffect, useCallback } from "react";
-import { useSelector } from "react-redux";
 import styled from "styled-components";
 import { Input } from "components/input/input.component";
 import { useMusic } from "lib/music/hook";
-import { useFirebase } from "lib/firebase/hook";
 import { Song } from "lib/music/types";
 import { MediaCard } from "components/media-card/media-card.component";
 import { useDebouncedCallback } from "hooks/use-debounced-callback";
-import { RootState } from "store/reducers";
-import { useParams } from "react-router-dom";
-import { MediaQueue, MediaQueueItem } from "../../../../components/media-queue";
+import { MediaQueue, MediaQueueItem } from "components/media-queue";
+import { useEnqueueSongFirebase } from "hooks/use-enqueue-song-firebase";
 
 const SEARCH_DEBOUNCE = 750; // Milliseconds
 
@@ -17,9 +14,7 @@ export const Search = () => {
   const [searchQuery, setSearchQuery] = useState("");
   const [searchResults, setSearchResults] = useState<Song[]>([]);
   const music = useMusic();
-  const firebase = useFirebase();
-  const userId = useSelector((state: RootState) => state.userId);
-  const { roomKey } = useParams<{ roomKey: string }>();
+  const enqueueSongFirebase = useEnqueueSongFirebase();
 
   const onSearch = useDebouncedCallback(
     () => {
@@ -37,9 +32,9 @@ export const Search = () => {
 
   const onClickAddSong = useCallback(
     (song: Song) => {
-      firebase.database().ref(`rooms/${roomKey}/queues/${userId}`).push(song);
+      enqueueSongFirebase(song);
     },
-    [firebase, roomKey, userId]
+    [enqueueSongFirebase]
   );
 
   return (
